Fix a compiler warning
git-svn-id: https://www.opensc-project.org/svnp/opensc/branches/martin/0.12@3838 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/pkcs11/sc-pkcs11.h
@@ -67,7 +67,7 @@ struct sc_pkcs11_card;
 
 /* Object Pool */
 struct sc_pkcs11_pool_item {
-	int handle;
+	unsigned long int handle;
 	void *item;
 	struct sc_pkcs11_pool_item *next;
 	struct sc_pkcs11_pool_item *prev;
